Jesse changed for:
We've detected a few false triggers do to some issues that I've known about,
but wasn't concerned about because we weren't getting as many triggers.  Now
that the sensors are more sensitive, we have several things to potentially
change.

I updated jfl_trigmon with a few new things.  I made sure to run an svn
update before starting in on it.  It seems to compile.  I would like these
new changes included in the running version of trigmon.

Major changes:
1) I think I fixed a bug where additional triggers from the same sensor can
be included by accident.
2) The sensors cannot be correlated if they come from within 10 meters of
each other (they are in the same building).
3) I use 4 sec rather than 5 sec between iterations.

Let me know if I need to update a different version, or are you still using
the jfl_trigmon.cpp source code.  If not, I'll be happy to make updates to
others.